tui/requirements.txt
textual==0.45.1
rich==13.7.0
pyyaml==6.0.1
click==8.1.7
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
textual-dev==1.2.1
